feat: implement pagination for user retrieval and add hashed token verification to AccessTokenGuard

+++ src/users/users.service.ts
@@ -1,8 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, isValidObjectId } from 'mongoose';
 import { User } from './schema/user.schema';
 import { UpdateUserDto } from '../auth/dto';
+import { QueryUserDto } from './dto/query-user.dto';
 
 @Injectable()
 export class UsersService {
@@ -11,8 +16,18 @@ export class UsersService {
     private readonly userModel: Model<User>,
   ) {}
   // ================= Find All =================
-  async findAll(): Promise<User[]> {
-    return this.userModel.find().select('-password').lean();
+  async findAll(query: QueryUserDto): Promise<{ data: User[]; total: number }> {
+    const { limit = 10, page = 1 } = query;
+    const skip = (page - 1) * limit;
+    const [data, total] = await Promise.all([
+      this.userModel.find().limit(limit).skip(skip).select('-password').lean(),
+      this.userModel.countDocuments(),
+    ]);
+
+    return {
+      total,
+      data,
+    };
   }
 
   // ================= Find One =================
@@ -22,14 +37,20 @@ export class UsersService {
 
   // ================= Update =================
   async update(id: string, dto: UpdateUserDto) {
-    const user = await this.userModel.findByIdAndUpdate(id, dto, { new: true });
-
+    const user = await this.userModel
+      .findByIdAndUpdate(id, { $set: dto }, { new: true, runValidators: true })
+      .select('-password');
     if (!user) throw new NotFoundException('User not found');
-
-    return { message: 'User updated successfully' };
+    return {
+      message: 'User updated successfully',
+      user,
+    };
   }
   // ================= Remove =================
   async remove(id: string) {
+    if (!isValidObjectId(id)) {
+      throw new BadRequestException('Invalid User ID format');
+    }
     const user = await this.userModel.findByIdAndDelete(id);
     if (!user) throw new NotFoundException('User not found');
     return { message: 'User deleted successfully' };
